+++ test/syc/good/gt.c
@@ -6,5 +6,5 @@ int a, b;
 
 int gt(void)
 {
-	return a > b;
+	return (int)(a > b);
 }
